project code/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Flask-Limiter
Flask-Mail
PyMySQL
Flask-Bcrypt
python-dotenv
pycryptodome
Pillow
Redis
